# docs/swagger.yaml
basePath: /
definitions:
  main.CreateTaskRequest:
    properties:
      status:
        type: string
      title:
        type: string
    type: object
  main.ListTasksResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/main.Task'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  main.Task:
    properties:
      created_at:
        type: string
      id:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
  main.UpdateTaskRequest:
    properties:
      status:
        type: string
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API untuk manajemen task (CRUD + pagination + filter status)
  title: Task Management API
  version: "1.0"
paths:
  /tasks:
    get:
      consumes:
      - application/json
      description: Ambil list task dengan pagination dan filter status
      parameters:
      - description: 'Filter by status (contoh: todo|done)'
        in: query
        name: status
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ListTasksResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List tasks
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Membuat task baru
      parameters:
      - description: Task payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create task
      tags:
      - Tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Hapus task berdasarkan ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete task
      tags:
      - Tasks
    get:
      consumes:
      - application/json
      description: Ambil detail task berdasarkan ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Task'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get task by ID
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Update task berdasarkan ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update task
      tags:
      - Tasks
swagger: "2.0"